Removing the chmod of the mongo-data directory as well.
I removed the rm earlier. The chmod fails too. Duh.

We do still need to figure out how to unravel the ownership
inconsistency here - maybe the ideal is really just to get this
running on a box with a big enough root partition (or whatever
partition has /var/lib/docker on it) to not need us to move the data
directory to a bind mounted volume. *shrug*