Made robust against data fudging in the source
When we discover issues in the data source, we sometimes need to reprocess the old data, as it happened with INFRA-896.
When that happens, corresponding files would also get deleted from census host.

This change makes the processing robust against that kind of data manipulation. Without the --delete flag, stale copy in the workspace can cause some data to get double-counetd.