Only process last month
I don't understand why I put that mess with "last3Months" previously,
but what's sure is that it currently generates wrong data. Around the
sum of the last 2 months, instead of just the last one...

So this fix only looks at the last month found, as do the other data
analysis and as one would expect...

Just tested with the data I have locally (December 2015), and I now find
2733 (instead of 2737) for buildtriggerbadge, or 124615 (instead 124626)
for maven-plugin, but it's normal.

This difference comes from the additional

    "and version NOT LIKE '%(private)' "

only put here because those private are going to be shown. But they're
kept in the latestNumbers count since their names are not disclosed
anyway there.